health to 2 bytes

+++ Common/Snapshot/OperatorDto.cpp
@@ -1,7 +1,7 @@
 #include "OperatorDto.h"
 
 StOperator::StOperator(const uint8_t& id, const TypeOperator& type, const State& state,
-        const std::pair<int16_t, int16_t>& position, const uint8_t& health, const uint8_t& munition,
+        const std::pair<int16_t, int16_t>& position, const int16_t& health, const uint8_t& munition,
         const bool& grenadeAvailable, const bool& smokeAvailable) :
         id(id), type(type), state(state), position(position), health(health), munition(munition),
         grenadeAvailable(grenadeAvailable), smokeAvailable(smokeAvailable) {}
@@ -22,7 +22,7 @@ std::pair<int16_t, int16_t> StOperator::getPosition() const {
     return position;
 }
 
-uint8_t StOperator::getHealth() const {
+int16_t StOperator::getHealth() const {
     return health;
 }
